backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
